# Optimize `IntRange::from_pat`, then shrink `ParamEnv`