@vandenheuvel 2020-09-28
Very interesting. Please let me know if I'm understanding your comment correctly: combining these two issues results in this PR having a net gain, but this doesn't address the issue underlying #76913? So this would not close https://github.com/rust-lang/rust/issues/77058?