@ecstatic-morse 2020-09-28
I'm not quite sure what you mean @simonvandel. I think I've gone as far as I care to. If you'd like to investigate further, feel free to keep #77058 open. Perhaps you can do even better. 

Here's the "slow" assembly (`ParamEnv` shrunk with no other changes) in `rustc_mir_build::thir::pattern::_match::PatStack::specialize_constructor()` annotated via `perf` with the sum of `StoreQueueRsrcStall` and `LoadQueueRsrcStall` perf events (I'm on a Zen 2 Ryzen processor).

`IntRange::from_pat` (inlined)
```asm
  0.00        → callq  rustc_mir_build::thir::pattern::_match::pat_constructor
  0.00          mov    0xb0(%rsp),%cl
  0.02          mov    0xc0(%rsp),%r9
  0.20          mov    0xb8(%rsp),%rbp
  0.01          mov    0xd0(%rsp),%r13
  0.01          mov    0xc8(%rsp),%r8
  0.01          mov    0xd8(%rsp),%al
  0.01          mov    0xd9(%rsp),%edx
 32.37          mov    %edx,0x40(%rsp)
  0.00          mov    0xdc(%rsp),%edx
  0.00          mov    %edx,0x43(%rsp)
                cmp    $0x7,%cl 
              ↓ jne    5a1      
```

`IntRange::intersection` (inlined)
```asm
         5a1:   mov    0x40(%rsp),%edx
 20.80          mov    0x43(%rsp),%esi
  0.03          mov    %esi,0x183(%rsp)
                mov    %edx,0x180(%rsp)
                mov    $0x1,%dl 
                cmp    $0x3,%cl 
              ↓ jne    5d9      
                mov    0x180(%rsp),%ecx
  0.01          mov    0x183(%rsp),%edx
 23.64          mov    %edx,0x23(%rsp)
  0.12          mov    %ecx,0x20(%rsp)
                cmp    $0x2,%al 
                sete   %dl      
  0.07   5d9:   mov    0x20(%rsp),%eax
  0.01          mov    0x23(%rsp),%ecx
 18.20          mov    %ecx,0xb3(%rsp)
  0.01          mov    %eax,0xb0(%rsp)
                test   %dl,%dl  
  0.00        ↓ jne    1546     
                mov    %r14,0x18(%rsp)
  0.00          mov    0x218(%rsp),%rax
  0.00          mov    (%rax),%rdi
  0.01          mov    0x30(%rbx),%rax
  0.01          mov    0x8(%rbx),%r14
  0.14          mov    0x10(%rbx),%rsi
                mov    0x20(%rbx),%rdx
                mov    0x18(%rbx),%r10
                movzbl (%rax),%ecx
                cmp    $0x2,%rcx
              ↓ je     629      
```

We're repeatedly copying two words into local stack variables, then branching, then reading them again. This sequence has a lot of data hazards, which is clearly slowing down the pipeline significantly. I can't say why LLVM chose to generate code this way, or what causes the data hazards to disappear with my changes to `IntRange::from_pat`. You'd need someone with more LLVM expertise than I have. Perhaps building with `debug-info-level=2` and stepping through it with a debugger would be helpful?